+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/finder/PlaceholdersEnum.groovy
@@ -1,0 +1,12 @@
+package ru.vyarus.guice.persist.orient.support.finder;
+
+/**
+ * Placeholder definition with enum.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 22.09.2014
+ */
+public enum PlaceholdersEnum {
+    name,
+    nick
+}