+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/modules/TestFinderModule.groovy
@@ -2,6 +2,7 @@ package ru.vyarus.guice.persist.orient.support.modules
 
 import com.google.inject.AbstractModule
 import ru.vyarus.guice.persist.orient.FinderModule
+import ru.vyarus.guice.persist.orient.support.finder.FinderWithPlaceholders
 import ru.vyarus.guice.persist.orient.support.finder.InterfaceFinder
 
 /**
@@ -13,6 +14,6 @@ class TestFinderModule extends AbstractModule {
     @Override
     protected void configure() {
         install(new PackageSchemeModule())
-        install(new FinderModule(InterfaceFinder))
+        install(new FinderModule(InterfaceFinder, FinderWithPlaceholders))
     }
 }