+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 * Fix finder module available db types detection
 * Add query placeholders support for finders
+* Update orient (1.7.8 > 1.7.9) - [important hotfix](https://groups.google.com/forum/#!topic/orient-database/vPF85I5Blts)
+
 
 ### 1.0.3 (2014-09-15)